requirements.txt
Django==1.11
djangorestframework==3.9.3
# alabaster==0.7.12
# anaconda-client==1.7.2
# anaconda-navigator==1.9.7
# anaconda-project==0.8.3
# asn1crypto==1.0.1
# astroid==1.6.5
# astropy==2.0.9
# atomicwrites==1.3.0
# attrs==19.2.0
# audioread==2.1.9
# Babel==2.7.0
# backports-abc==0.5
# backports.functools-lru-cache==1.5
# backports.os==0.1.1
# backports.shutil-get-terminal-size==1.0.0
# backports.ssl-match-hostname==3.7.0.1
# backports.tempfile==1.0
# backports.weakref==1.0.post1
# beautifulsoup4==4.8.0
# bitarray==1.0.1
# bkcharts==0.2
# bleach==3.1.0
# bokeh==1.3.4
# boto==2.49.0
# Bottleneck==1.2.1
# cdecimal==2.3
# certifi==2019.9.11
# cffi==1.12.3
# chardet==3.0.4
# Click==7.0
# cloudpickle==1.2.2
# clyent==1.2.2
# colorama==0.4.1
# conda==4.7.12
# conda-build==3.18.9
# conda-package-handling==1.6.0
# conda-verify==3.4.2
# configparser==4.0.2
# contextlib2==0.6.0
# cryptography==2.7
# cycler==0.10.0
# Cython==0.29.13
# cytoolz==0.10.0
# dask==1.2.2
# deap==1.3.1
# decorator==4.4.0
# defusedxml==0.6.0
# dill==0.3.4
# distributed==1.28.1
# docutils==0.15.2
# entrypoints==0.3
# enum34==1.1.6
# essentia==2.1b6.dev234
# et-xmlfile==1.0.1
# eyeD3==0.8.12
# fastcache==1.1.0
# filelock==3.0.12
# Flask==1.1.1
# funcsigs==1.0.2
# functools32==3.2.3.post2
# future==0.17.1
# futures==3.3.0
# gevent==1.4.0
# glob2==0.7
# gmpy2==2.0.8
# greenlet==0.4.15
# grin==1.2.1
# h5py==2.9.0
# HeapDict==1.0.1
# html5lib==1.0.1
# idna==2.8
# imageio==2.6.0
# imagesize==1.1.0
# importlib-metadata==0.23
# ipaddress==1.0.22
# ipykernel==4.10.0
# ipython==5.8.0
# ipython-genutils==0.2.0
# ipywidgets==7.5.1
# isort==4.3.21
# itsdangerous==1.1.0
# jdcal==1.4.1
# jedi==0.15.1
# Jinja2==2.10.3
# joblib==0.14.1
# jsonschema==3.0.2
# jupyter==1.0.0
# jupyter-client==5.3.3
# jupyter-console==5.2.0
# jupyter-core==4.5.0
# jupyterlab==0.33.11
# jupyterlab-launcher==0.11.2
# kiwisolver==1.1.0
# lazy-object-proxy==1.4.2
# libarchive-c==2.8
# librosa==0.7.1
# lief==0.9.0
# linecache2==1.0.0
# llvmlite==0.29.0
# locket==0.2.0
# lxml==4.4.1
# MarkupSafe==1.1.1
# matplotlib==2.2.3
# mccabe==0.6.1
# mistune==0.8.4
# mkl-fft==1.0.14
# mkl-random==1.1.0
# mkl-service==2.3.0
# mock==3.0.5
# more-itertools==5.0.0
# mpmath==1.1.0
# msgpack==0.6.1
# multipledispatch==0.6.0
# multiprocess==0.70.12.2
# navigator-updater==0.2.1
# nbconvert==5.6.0
# nbformat==4.4.0
# networkx==2.2
# nltk==3.4.5
# nose==1.3.7
# notebook==5.7.8
# numba==0.45.1+0.ga9c107beb.dirty
# numexpr==2.7.0
# numpy==1.16.5
# numpydoc==0.9.1
# olefile==0.46
# openpyxl==2.6.3
# packaging==19.2
# pandas==0.24.2
# pandocfilters==1.4.2
# parso==0.5.1
# partd==1.0.0
# path.py==11.5.0
# pathlib==1.0.1
# pathlib2==2.3.5
# pathos==0.2.8
# patsy==0.5.1
# pep8==1.7.1
# pexpect==4.7.0
# pickleshare==0.7.5
# Pillow==6.2.0
# pkginfo==1.5.0.1
# pluggy==0.13.0
# ply==3.11
# pox==0.3.0
# ppft==1.6.6.4
# prometheus-client==0.7.1
# prompt-toolkit==1.0.15
# psutil==5.6.3
# ptyprocess==0.6.0
# py==1.8.0
# pyAudioAnalysis==0.2.5
# pycairo==1.18.1
# pycodestyle==2.5.0
# pycosat==0.6.3
# pycparser==2.19
# pycrypto==2.6.1
# pycurl==7.43.0.3
# pydotplus==2.0.2
# pydub==0.23.1
# pyflakes==2.1.1
# Pygments==2.4.2
# pylint==1.9.2
# pyodbc==4.0.27
# pyOpenSSL==19.0.0
# pyparsing==2.4.2
# pyrsistent==0.15.4
# PySocks==1.7.1
# pytest==4.6.2
# python-dateutil==2.8.0
# python-magic==0.4.24
# pytz==2019.3
# PyWavelets==1.0.3
# PyYAML==5.1.2
# pyzmq==18.1.0
# QtAwesome==0.6.0
# qtconsole==4.5.5
# QtPy==1.9.0
requests==2.22.0
# resampy==0.2.2
# rope==0.14.0
# ruamel-yaml==0.15.46
# scandir==1.10.0
# scikit-image==0.14.2
# scikit-learn==0.20.3
# scipy==1.2.1
# seaborn==0.9.0
# Send2Trash==1.5.0
# simplegeneric==0.8.1
# singledispatch==3.4.0.3
# six==1.12.0
# snowballstemmer==2.0.0
# sortedcollections==1.1.2
# sortedcontainers==2.1.0
# SoundFile==0.10.3.post1
# soupsieve==1.9.3
# Sphinx==1.8.5
# sphinxcontrib-websupport==1.1.2
# spyder==3.3.6
# spyder-kernels==0.5.2
# SQLAlchemy==1.3.9
# statsmodels==0.10.1
# subprocess32==3.5.4
# sympy==1.4
# tables==3.5.2
# tblib==1.4.0
# terminado==0.8.2
# testpath==0.4.2
# toolz==0.10.0
# tornado==5.1.1
# tqdm==4.36.1
# traceback2==1.4.0
# traitlets==4.3.3
# typing==3.7.4.1
# unicodecsv==0.14.1
# unittest2==1.1.0
# urllib3==1.24.2
# wcwidth==0.1.7
# webencodings==0.5.1
# Werkzeug==0.16.0
# widgetsnbextension==3.5.1
# wrapt==1.11.2
# wurlitzer==1.0.3
# xgboost==0.82
# xlrd==1.2.0
# XlsxWriter==1.2.1
# xlwt==1.3.0
# zict==1.0.0
# zipp==0.6.0